Local write and read index works. TODO: work on download lambda.
